Move markExtend into selections
Needed by tests, and best not to contaminate module.exports of commands

Alternatively this module could be called `utils`
